+++ app/src/main/java/com/github/sdp/ratemyepfl/database/RoomNoiseRepository.kt
@@ -7,7 +7,7 @@ import java.time.LocalDateTime
 
 interface RoomNoiseRepository {
 
-    suspend fun addMeasurement(roomId: String, date: LocalDateTime, measure: Double): Task<Transaction>
+    suspend fun addMeasurement(roomId: String, date: LocalDateTime, measure: Int): Task<Transaction>
 
     suspend fun getRoomNoiseInfoById(roomId: String): RoomNoiseInfo?
 }
